Refactor: separate module for utilities related to loading/saving checkpoints and safetensors
